Make all records safely live in a single data stream
